Heart Monitoring Application
This android application allows user to put in the patient details and
when the run button is hit the application accesses the accelerometer
of the smartphone to get the x,y,z axis configuration and plots the
heart rate of the patient in a continuously moving graph. The data from
the accelerometer is also uploaded to a web server database. This data
can also be downloaded from the application directly to your phone.